create local lib to do network
