+++ 5sem/3d_block/proxy/Makefile
@@ -2,7 +2,7 @@ CC = gcc
 FLAGS = -Wall -g
 LIBS = -lpthread
 TARGET = cache-proxy
-COMMON = http-parse.o
+COMMON = http-parse.o network.o
