+++ 5sem/3d_block/proxy/network.h
@@ -1,0 +1,3 @@
+int create_server_socket_and_listen(int port);
+
+int connect_to_remote_server(struct hostent* server);
